Changed how <text/> is reported.
Previous versions of salve would indicate that ``<text/>`` was possible
by returning an event with ``"text"`` as the first parameter and ``"*"``
as the second. This was ambiguous because a ``<value>`` that allows only
an asterisk would also return the same event. ``<text/>`` is now
indicated by the regular expression ``/^.*$/`` in the second position.