Fix an inclusion problem.
Fixed an error whereby if an RNG file processed with ``salve-convert`` included
another RNG file, and the other file had a ``datatypeLibrary`` set on the top
``grammar`` element, this attribute would be lost during the schema
simplification process. One symptom would be to cause any reference to types
that are not in the default library to fail during conversion.