Use mergeOptions rather than _.clone. Don't depend on lodash.
The clones were shallow and getting everythig to clone properly was very error
prone. It would have been easy to eliminate tests by accident. Using
mergeOptions does a deep copy and is thus less likely to allow for errors.

With a few additional modifications. This has the side-effect that we can drop
lodash as a dependency.