Don't delay walker creation.
We previously were delaying the creation of subwalkers but that was not a useful
optimization because in normal usage the subwalkers are pretty much always
created anyway. Conversely, it had a cost in additional code and run time.

Historical note: delaying the creation of subwalkers was helpful in reducing
some scenarios dealing with cloning walkers, but cloning should be relatively
infrequent.