Segregate the ElementWalker objects.
Salve used to simply send events through a tree of walkers. That method of doing
things complicated the logic of walkers because we could not take advantage of
the logical separation provided by entering an element (e.g. ``enterStartTag``
events can only pertain to the element we are currently in, so there's no point
is having parent elements see the event.)

We now segregate the trees by implementing a stack of ElementWalker objects on
GrammarWalker. This way, the logic internal to the various walkers can be
simplified and the code can run faster.