Update canEnd status as we go.
Salve now maintains two flags canEnd and canEndAttribute which are updated as
events are fired, instead of having a canEnd function which computes these when
called. The fact of the matter is that checking whether a walker can end will
happen most of the time, so there's no point in delaying the check.