refact: drop the autoload modules in favor of requireDir
The autoload modules were rather awkward. They had an effect only in Node but
had to pass through Webpack, etc. We've removed them in favor of requireDir. Our
single script uses requireDir so that all possible validators and simplifiers
are loaded in Node. In the browser we support only the internal
validator/simplifier, which we load directly.