build: bootstrap without having to use an older salve install
The build process now makes use of the new "none" validator to bootstrap the
conversion of ``relaxng.rng`` into a JSON file. This allows not having to depend
on an older installed salve, and makes it so that any improvement to the
simplification/conversion make it immediately into the ``relaxng.json`` bundled
with the next release. (Otherwise, it would be one release behind.)