refact: compile to ES6
This change also removes the polyfills we were using as they are no longer of
any use after dropping IE11 and Safari 9.

BREAKING CHANGE:

This removes support for IE11 and Safari 9. The change is justified by the fact
that we get a 10% speed improvement "for free" by compiling to ES6 when running
the docbook.rng benchmark. Worth it. If there is demand for IE11 or Safari 9
support, we may add a build for those platforms. If there is demand.