build: use TRAVIS_PULL_REQUEST_BRANCH
When Travis makes a "pr" build, the build is made by merging the pr with the
main branch and running a build on that. In that case TRAVIS_BRANCH is set to
the main branch's name instead of the PR's branch name, and it messes up our
logic. The change fixes the problem. We cannot just use
TRAVIS_PULL_REQUEST_BRANCH all the time because on for the "push" job which is
launched, TRAVIS_PULL_REQUEST_BRANCH is not set.