feat: simplification can now produce a manifest
``convertRNGToPattern`` by default will skip producing a manifest but if called
with ``createManifest`` set to ``true`` it will create a manifest of the files
it had to read and the sha512 checksum of each file.

BREAKING CHANGE:

The support for manifests requires using Node functions that are only available
in Node 8 and higher. So this means Node 6 is no longer supported. If you need
Node 6 support, you can use salve 7.0.3 or earlier. Alternatively, we may
consider a PR to reintroduce support for Node 6. If you do produce a PR make
sure that your code does run in Node 6 and that the PR is polished. A PR that
does the job half-way won't be accepted. Just opening an issue asking for Node 6
support is unlikely to go anywhere.