feat: GrammarWalker now takes an external name resolver
BREAKING CHANGE:

``GrammarWalker`` now takes an external name resolver. Prior to this change,
``GrammarWalker`` would instantiate an instance of ``NameResolver``. This was
disadvantageous in cases where the code using salve already maintained its own
name resolving machinery. It would essentially have to duplicate information
with salve and suffer the performance consequences.

``NameResolver`` has been renamed to ``DefaultNameResolver`` and
``NameResolver`` is now an interface. When calling ``newWalker`` on ``Grammar``
you now need to pass a ``NameResolver`` instance. For cases where the old
behavior was as good as it gets, the way to get your code working with a minimum
of changes is to: a) replace ``const walker = grammar.newWalker()`` with ``const
walker = grammar.newWalker(new DefaultNameResolver())``, b) instead of calling
the ``DefaultNameResolver`` methods ``enterContext``, ``definePrefix``,
etc. directly on the ``walker``, call them on ``walker.nameResolver``.