build: move git-tag-config from package.json to npmrc
It was useless in `package.json` because those configuration values are used for
scripts, not for setting configuration values for npm commands. So when `npm
version`'s code checks configuration, it is not affected by the `config` section
in `package.json`. The configuration set in `package.json` *is* however
influencing what configuration the `{pre,post,}version` scripts see.