perf: simplify _prepare
_prepare would gather the references (Ref) that cannot be resolved and would
produce a final and nice error message. The fact of the matter is that if a
reference cannot be resolved when _prepare is called, there's been an internal
error somewhere else. So instead of trying to be nice, abort loudly at the first
sign of trouble. Diagnosis will have to be performed through other means.

BREAKING CHANGE: RefError is no longer exported.