perf: check invalid paths in <start> separately
Once upon a time we were querying the tree for the invalid paths. Then we
switched to keeping state while we walk the tree. The issue with state is that
for all the paths that are invalid in `<start>` we did the path tests over the
whole tree. It is better to query the problematic elements and check for invalid
paths in `<start>` in one shot.